Remove fastify test as it is not supported for iast and @apollo/server test
