+++ packages/dd-trace/test/appsec/iast/taint-tracking/sources/graphql.sources.test-utils.js
@@ -25,12 +25,12 @@ query GetBooks ($title: String) {
 }`
 
 const queryWithHardcodedArgument = `
-    query GetBooks {
-      books(title: "ls") {
-        title,
-        author
-      }
-    }`
+query GetBooks {
+  books(title: "ls") {
+    title,
+    author
+  }
+}`
 
 const books = [
   {
